dont try to delete pending agencies if they are on another request

+++ muckrock/foia/views/composers.py
@@ -378,9 +378,12 @@ def autosave(request, idx):
         composer = form.save()
         new_agencies = set(composer.agencies.all())
         removed_agencies = old_agencies - new_agencies
-        # delete pending agencies which have been removed from all composers
+        # delete pending agencies which have been removed from composers and requests
         for agency in removed_agencies:
-            if agency.status == 'pending' and agency.composers.count() == 0:
+            if (
+                agency.status == 'pending' and agency.composers.count() == 0
+                and agency.foiarequest_set.count() == 0
+            ):
                 agency.delete()
         return HttpResponse('OK')
     else:
